# sdk/openapi/agent2allow.openapi.json
{
  "components": {
    "schemas": {
      "ApprovalDecisionRequest": {
        "properties": {
          "approver": {
            "default": "human",
            "title": "Approver",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "ApprovalDecisionRequest",
        "type": "object"
      },
      "ApprovalView": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "repo": {
            "title": "Repo",
            "type": "string"
          },
          "request_payload": {
            "additionalProperties": true,
            "title": "Request Payload",
            "type": "object"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "tool",
          "action",
          "repo",
          "risk_level",
          "request_payload",
          "reason",
          "created_at",
          "updated_at"
        ],
        "title": "ApprovalView",
        "type": "object"
      },
      "AuditLogView": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "repo": {
            "title": "Repo",
            "type": "string"
          },
          "request_payload": {
            "additionalProperties": true,
            "title": "Request Payload",
            "type": "object"
          },
          "response_payload": {
            "additionalProperties": true,
            "title": "Response Payload",
            "type": "object"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "id",
          "timestamp",
          "agent_id",
          "tool",
          "action",
          "repo",
          "risk_level",
          "schema_version",
          "status",
          "request_payload",
          "response_payload",
          "approval_id",
          "message"
        ],
        "title": "AuditLogView",
        "type": "object"
      },
      "BulkApprovalRequest": {
        "properties": {
          "approver": {
            "default": "human",
            "title": "Approver",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "ids": {
            "items": {
              "type": "integer"
            },
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "ids",
          "decision"
        ],
        "title": "BulkApprovalRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ToolCallRequest": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "agent_id": {
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "repo": {
            "title": "Repo",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "tool",
          "action",
          "repo"
        ],
        "title": "ToolCallRequest",
        "type": "object"
      },
      "ToolCallResponse": {
        "properties": {
          "approval_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "idempotent_replay": {
            "default": false,
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "ToolCallResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Agent2Allow",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/ready": {
      "get": {
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "boolean"
                      },
                      {
                        "additionalProperties": {
                          "type": "boolean"
                        },
                        "type": "object"
                      }
                    ]
                  },
                  "title": "Response Ready Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready"
      }
    },
    "/v1/approvals/bulk": {
      "post": {
        "operationId": "approvals_bulk_v1_approvals_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approvals Bulk V1 Approvals Bulk Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approvals Bulk"
      }
    },
    "/v1/approvals/pending": {
      "get": {
        "operationId": "approvals_pending_v1_approvals_pending_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalView"
                  },
                  "title": "Response Approvals Pending V1 Approvals Pending Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Approvals Pending"
      }
    },
    "/v1/approvals/{approval_id}/approve": {
      "post": {
        "operationId": "approvals_approve_v1_approvals__approval_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approvals Approve V1 Approvals  Approval Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approvals Approve"
      }
    },
    "/v1/approvals/{approval_id}/deny": {
      "post": {
        "operationId": "approvals_deny_v1_approvals__approval_id__deny_post",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approvals Deny V1 Approvals  Approval Id  Deny Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approvals Deny"
      }
    },
    "/v1/audit": {
      "get": {
        "operationId": "audit_logs_v1_audit_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditLogView"
                  },
                  "title": "Response Audit Logs V1 Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Audit Logs"
      }
    },
    "/v1/audit/export": {
      "get": {
        "operationId": "audit_export_v1_audit_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Audit Export V1 Audit Export Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Audit Export"
      }
    },
    "/v1/tool-calls": {
      "post": {
        "operationId": "tool_calls_v1_tool_calls_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tool Calls"
      }
    }
  }
}